% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{amd_hdl}
\alias{amd_hdl}
\title{Effect of HDL choesterol on age-related macular degeneration (AMD)}
\format{
A \code{data.frame} with 27 rows and 4 variables.
}
\usage{
data(amd_hdl)
}
\description{
Summary data obtained by combining the following genome-wide association studies:
\enumerate{
\item{SNP-HDL:}{Kettunen et al. Available from: http://europepmc.org/articles/PMC4814583}
\item{SNP-AMD:}{International AMD Genomics Consortium (IAMDGC) Available from: http://europepmc.org/articles/PMC4745342}
}
}
\details{
The Metabolic Syndrome in Men (METSIM) GWAS has been used for SNP selection.
}
\keyword{datasets}
