% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_fit_list-class}
\alias{bayesmr_fit_list-class}
\alias{bayesmr_fit_list}
\title{List of fitted BayesMR fixed-heterogeneity chains.}
\description{
Stores the output from one or more MCMC chains for a BayesMR model without
mixture clustering and with fixed heterogeneity.
}
\section{Slots}{

\describe{
\item{\code{results}}{A list of \code{\link[=bayesmr_fit-class]{bayesmr_fit}} objects, one for each simulated
chain.}
}}

\references{
Consonni, G., Venturini, S., Castelletti, F. (2026).
"Bayesian Hierarchical Modeling for Two-Sample Summary-Data Mendelian
Randomization under Heterogeneity". Working paper.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
