% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_fit_list_to_list}
\alias{bayesmr_fit_list_to_list}
\title{Conversion of a \code{bayesmr_fit_list} Object to a \code{list}}
\usage{
bayesmr_fit_list_to_list(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of class \code{bayesmr_fit_list}.}

\item{include.burnin}{Logical; if \code{TRUE} the burn-in iterations
(when stored) are included. Default is \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE} prints warnings during conversion.
Default is \code{TRUE}.}
}
\value{
A list of numeric matrices, one per chain, with columns
\code{gamma}, \code{beta}, \code{loglik}, \code{logprior},
\code{logpost}.
}
\description{
Converts a multi-chain \code{\link{bayesmr_fit_list-class}} object to a
plain list of matrices (one per chain) suitable for use with
\pkg{bayesplot} functions.
}
\examples{
\dontrun{
data(ldl_cad)
dat <- new("bayesmr_data", ldl_cad)
res <- bayesmr(dat, control = bayesmr_control(nsim = 500, burnin = 500,
                                              nchains = 2))
lst <- bayesmr_fit_list_to_list(res)
bayesplot::mcmc_trace(lst, pars = "gamma")
}

}
\seealso{
\code{\link{bayesmr}()}, \code{\link{bayesmr_fit_list-class}},
\code{\link{bayesmr_fit_list_to_mcmc.list}()}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
