% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_fit_list_to_mcmc.list}
\alias{bayesmr_fit_list_to_mcmc.list}
\title{Conversion of a \code{bayesmr_fit_list} Object to \code{mcmc.list}}
\usage{
bayesmr_fit_list_to_mcmc.list(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of class \code{bayesmr_fit_list}.}

\item{include.burnin}{Logical; if \code{TRUE} the burn-in iterations
(when stored) are included. Default is \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE} prints warnings during conversion.
Default is \code{TRUE}.}
}
\value{
A \code{\link[coda]{mcmc.list}} object.
}
\description{
Converts a multi-chain \code{\link{bayesmr_fit_list-class}} object to a
\code{\link[coda]{mcmc.list}} for convergence diagnostics.
}
\examples{
\dontrun{
data(ldl_cad)
dat <- new("bayesmr_data", ldl_cad)
res <- bayesmr(dat, control = bayesmr_control(nsim = 500, burnin = 500,
                                              nchains = 2))
mc <- bayesmr_fit_list_to_mcmc.list(res)
coda::gelman.diag(mc)
}

}
\seealso{
\code{\link{bayesmr}()}, \code{\link{bayesmr_fit_list-class}},
\code{\link[coda]{mcmc.list}}, \code{\link{bayesmr_fit_list_to_list}()}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
