% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_fit_to_mcmc}
\alias{bayesmr_fit_to_mcmc}
\title{Conversion of a \code{bayesmr_fit} Object to \code{mcmc}}
\usage{
bayesmr_fit_to_mcmc(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of class \code{bayesmr_fit}.}

\item{include.burnin}{Logical; if \code{TRUE} the burn-in iterations
(when stored) are included. Default is \code{FALSE}.}

\item{verbose}{Logical; if \code{TRUE} prints warnings during conversion.
Default is \code{TRUE}.}
}
\value{
A \code{\link[coda]{mcmc}} object.
}
\description{
Converts a single-chain \code{\link{bayesmr_fit-class}} object to a
\code{\link[coda]{mcmc}} object containing the \eqn{\gamma}, \eqn{\beta},
and log-density chains.
}
\examples{
\dontrun{
data(ldl_cad)
dat <- new("bayesmr_data", ldl_cad)
res <- bayesmr(dat, control = bayesmr_control(nsim = 500, burnin = 500))
mc <- bayesmr_fit_to_mcmc(res@results[[1]])
plot(mc)
}

}
\seealso{
\code{\link{bayesmr}()}, \code{\link{bayesmr_fit-class}},
\code{\link[coda]{mcmc}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
