% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_het_fit-class}
\alias{bayesmr_het_fit-class}
\title{Fitted BayesMR model with random heterogeneity.}
\description{
Stores the output from a single MCMC chain for a BayesMR model without mixture
clustering and with random heterogeneity.
}
\section{Slots}{

\describe{
\item{\code{gamma.chain}}{An array containing posterior draws of \code{gamma}.}

\item{\code{beta.chain}}{An array containing posterior draws of \code{beta}.}

\item{\code{psi.chain}}{An array containing posterior draws of \code{psi}.}

\item{\code{tau.chain}}{An array containing posterior draws of \code{tau}.}

\item{\code{accept}}{A matrix containing Metropolis-Hastings acceptance information.}

\item{\code{data}}{A data frame containing the analyzed summary data.}

\item{\code{dens}}{A list with log-density components, typically \code{loglik}, \code{logprior},
and \code{logpost}.}

\item{\code{control}}{A named list of MCMC control settings.}

\item{\code{prior}}{A named list of prior hyperparameters.}

\item{\code{dim}}{A list of model dimensions, including \code{n}, the number of SNPs.}
}}

\references{
Consonni, G., Venturini, S., Castelletti, F. (2026).
"Bayesian Hierarchical Modeling for Two-Sample Summary-Data Mendelian
Randomization under Heterogeneity". Working paper.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
