% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_init.R
\name{bayesmr_init}
\alias{bayesmr_init}
\title{Generate Starting Values for BayesMR Model Fitting}
\usage{
bayesmr_init(data, random_start = TRUE, K_start = 1, start_values)
}
\arguments{
\item{data}{An object of class \code{\link{bayesmr_data-class}} containing
the SNP-level summary statistics.}

\item{random_start}{A length-one logical. If \code{TRUE} (default),
starting values are drawn randomly from diffuse distributions. If
\code{FALSE}, the values in \code{start_values} are validated and
returned.}

\item{K_start}{A length-one positive integer giving the initial number of
mixture clusters. Use \code{K_start = 1} for the non-mixture models
(\code{\link{bayesmr}()}, \code{\link{bayesmr_het}()}).}

\item{start_values}{A named list of user-supplied starting values, used
when \code{random_start = FALSE}. See \code{\link{bayesmr_startvalues}()}
for the expected format.}
}
\value{
A named list of starting values. For \code{K_start = 1}: contains
\code{gamma}, \code{beta}, \code{psi}, \code{tau}. For
\code{K_start > 1}: also contains \code{K}, \code{xi}, \code{alpha}.
}
\description{
\code{bayesmr_init()} draws or validates starting values for the MCMC
sampler before fitting a BayesMR model.
}
\examples{
\dontrun{
data(ldl_cad)
dat <- new("bayesmr_data", ldl_cad)
# Random start for the no-clustering model
sv1 <- bayesmr_init(dat, random_start = TRUE, K_start = 1)
# Random start for the mixture model with 3 initial clusters
sv3 <- bayesmr_init(dat, random_start = TRUE, K_start = 3)
}

}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2026), "Bayesian Hierarchical
Modeling for Two-Sample Summary-Data Mendelian Randomization under
Heterogeneity", working paper.
}
\seealso{
\code{\link{bayesmr}()}, \code{\link{bayesmr_startvalues}()},
\code{\link{bayesmr_data-class}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
