% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_fit.R
\name{bayesmr_logLik}
\alias{bayesmr_logLik}
\title{Log-likelihood for BayesMR models.}
\usage{
bayesmr_logLik(data, gammaj, Gammaj)
}
\arguments{
\item{data}{A data frame ...}

\item{gammaj}{A numeric vector ...}

\item{Gammaj}{A numeric vector ...}
}
\value{
A length-one numeric vector of the log-likelihood value.
}
\description{
\code{bayesmr_logLik()} computes the log-likelihood value for a BayesMR model.
}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2024), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\seealso{
\code{\link{bayesmr}()}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
