% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_mix_fit_list_to_list}
\alias{bayesmr_mix_fit_list_to_list}
\title{Conversion of a \code{bayesmr_mix_fit_list} object to a \code{list}.}
\usage{
bayesmr_mix_fit_list_to_list(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of type \code{bayesmr_mix_fit_list}.}

\item{include.burnin}{A logical scalar.}

\item{verbose}{A logical scalar.}
}
\value{
A list of matrices (one per chain).
}
\description{
Conversion of a \code{bayesmr_mix_fit_list} object to a \code{list}.
}
\seealso{
\code{\link{bayesmr_mix}()}, \code{\link{bayesmr_mix_fit_list-class}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
