% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_mix_het_fit_list_to_mcmc.list}
\alias{bayesmr_mix_het_fit_list_to_mcmc.list}
\title{Conversion of a \code{bayesmr_mix_het_fit_list} object to an object of class \code{mcmc.list}.}
\usage{
bayesmr_mix_het_fit_list_to_mcmc.list(
  res,
  include.burnin = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{res}{An object of type \code{bayesmr_mix_het_fit_list}.}

\item{include.burnin}{A logical scalar.}

\item{verbose}{A logical scalar.}
}
\value{
An object of type \code{mcmc.list}.
}
\description{
Conversion of a \code{bayesmr_mix_het_fit_list} object to an object of class \code{mcmc.list}.
}
\seealso{
\code{\link{bayesmr_mix_het}()}, \code{\link{bayesmr_mix_het_fit_list-class}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
