% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_mix_het_fit_to_mcmc}
\alias{bayesmr_mix_het_fit_to_mcmc}
\title{Conversion of a \code{bayesmr_mix_het_fit} object to an object of class \code{mcmc}.}
\usage{
bayesmr_mix_het_fit_to_mcmc(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of type \code{bayesmr_mix_het_fit}.}

\item{include.burnin}{A logical scalar.}

\item{verbose}{A logical scalar.}
}
\value{
An object of type \code{mcmc}.
}
\description{
\code{bayesmr_mix_het_fit_to_mcmc} converts an object of class
\code{bayesmr_mix_het_fit} to a \code{coda::mcmc} object containing
the scalar parameter chains (\eqn{\gamma}, \eqn{\alpha}, \eqn{\psi},
\eqn{\tau}) and log-density values.
}
\seealso{
\code{\link{bayesmr_mix_het}()}, \code{\link{bayesmr_mix_het_fit-class}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
