% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_model-class}
\alias{bayesmr_model-class}
\alias{bayesmr_model}
\title{An S4 class to represent a BayesMR model.}
\description{
An S4 class to represent a BayesMR model.
}
\section{Slots}{

\describe{
\item{\code{p}}{A length-one character vector representing the number of dimensions
of the latent space to use in the MDS analysis.}

\item{\code{G}}{A length-one numeric vector representing the number of clusters to
partition the subjects into.}

\item{\code{family}}{A length-one character vector representing the type of data to
analyze. Currently, it accepts only the 'binomial' value, but future
developments will include the possibility to analyze continuous,
multinomial and count data.}
}}

\examples{
showClass("bayesmr_model")

}
\references{
Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based
Clustering of Several Binary Dissimilarity Matrices: the \pkg{bayesmr}
Package in \code{R}", Journal of Statistical Software, 100, 16, 1--35, <10.18637/jss.v100.i16>.
}
