% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_optim.R
\name{bayesmr_noclus_optim}
\alias{bayesmr_noclus_optim}
\title{Alternating conditional maximization of the bivariate posterior for a BayesMR model with no clustering.}
\usage{
bayesmr_noclus_optim(
  data,
  prior,
  start = rep(0, 2),
  maxiter = 1000,
  tol = 1e-10,
  beta_min = -20,
  beta_max = 20,
  beta_step = 0.001,
  n = 1000,
  tol_x = 1e-10,
  tol_f = 1e-12,
  eps_small = 1e-08,
  verbose = TRUE
)
}
\arguments{
\item{data}{An object of class \code{\link{bayesmr_data}})..}

\item{prior}{A list containing the prior hyperparameters. See
\code{\link{bayesmr_prior}()} for more details.}

\item{start}{A named list of starting values for the MCMC algorithm (see
\code{\link{bayesmr_init}}).}

\item{p}{A length-one numeric vector indicating the number of dimensions of the
latent space.}

\item{G}{A length-one numeric vector indicating the number of cluster to
partition the \emph{S} subjects.}

\item{control}{A list of control parameters that affect the sampling
but do not affect the posterior distribution See
\code{\link{bayesmr_control}()} for more details.}
}
\value{
A \code{bayesmr_fit_list} object.
}
\description{
\code{bayesmr_fit()} is the main function that estimates a BayesMR model.
}
\examples{
\dontrun{
data(simdiss, package = "bayesmr")

G <- 3
p <- 2
prm.prop <- list(z = 1.5, alpha = .75)
burnin <- 20000
nsim <- 10000
seed <- 2301

set.seed(seed)

control <- list(burnin = burnin, nsim = nsim, z.prop = prm.prop[["z"]],
  alpha.prop = prm.prop[["alpha"]], random.start = TRUE, verbose = TRUE,
  nchains = 2, thin = 10, store.burnin = TRUE, threads = 2,
  parallel = "snow")
sim.bayesmr <- bayesmr(simdiss, p, G, control)

summary(sim.bayesmr, include.burnin = FALSE)

library(bayesplot)
library(ggplot2)
color_scheme_set("teal")
plot(sim.bayesmr, what = "trace", regex_pars = "eta")

z <- bayesmr_get_configuration(sim.bayesmr, chain = 1, est = "mean",
  labels = 1:16)
summary(z)
color_scheme_set("mix-pink-blue")
graph <- plot(z, size = 2, size_lbl = 3)
graph + panel_bg(fill = "gray90", color = NA)
}
}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2026), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\seealso{
\code{\link{bayesmr_data}} for a description of the data format.

\code{\link{bayesmr_fit_list}} for a description of the elements
included in the returned object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
