% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{compute_psm}
\alias{compute_psm}
\title{Compute the p × p Posterior Similarity Matrix}
\usage{
compute_psm(out)
}
\arguments{
\item{out}{Object from remove_burnin()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Symmetric numeric matrix, dim p × p, entries in [0,1]
}\if{html}{\out{</div>}}
}
\description{
\eqn{PSM_{i,j}} = proportion of iterations in which SNPs i and j
are assigned to the same cluster.
}
