% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_datasets.R
\docType{data}
\name{educ_alzheimer}
\alias{educ_alzheimer}
\title{Educational attainment → Alzheimer’s disease MR dataset}
\format{
A \code{data.frame} with 317 rows and 4 variables.
}
\source{
OpenGWAS (https://gwas.mrcieu.ac.uk/)
}
\usage{
data(educ_alzheimer)
}
\description{
Harmonised two-sample Mendelian randomization (MR) dataset
constructed using genome-wide significant SNPs for educational
attainment as instruments and summary statistics for
Alzheimer’s disease as the outcome.
}
\details{
Instruments were selected at genome-wide significance
(p < 5e-8) and LD-clumped (r2 = 0.001, 10,000 kb window).
Proxy SNPs (rsq >= 0.8) were allowed for the outcome GWAS.

EXPOSURE: Educational attainment

OpenGWAS ID: ieu-a-1239

Source study:
Lee JJ et al. (2018).
Gene discovery and polygenic prediction from a genome-wide
association study of educational attainment in 1.1 million individuals.
Nature Genetics, 50:1112–1121.

GWAS Catalog accession:
GCST006250

GWAS Catalog URL:
https://www.ebi.ac.uk/gwas/studies/GCST006250

OpenGWAS dataset page:
https://gwas.mrcieu.ac.uk/datasets/ieu-a-1239/

OUTCOME: Alzheimer’s disease

OpenGWAS ID: ebi-a-GCST90027158

Source study:
Kunkle BW et al. (2019).
Genetic meta-analysis of diagnosed Alzheimer’s disease
identifies new risk loci and implicates Aβ, tau,
immunity and lipid processing.
Nature Genetics, 51:414–430.

GWAS Catalog accession:
GCST90027158

GWAS Catalog URL:
https://www.ebi.ac.uk/gwas/studies/GCST90027158

OpenGWAS dataset page:
https://gwas.mrcieu.ac.uk/datasets/ebi-a-GCST90027158/

VARIABLES

\describe{
\item{beta_exposure}{Estimated SNP–education association.}
\item{se_exposure}{Standard error of SNP–education association.}
\item{beta_outcome}{Estimated SNP–Alzheimer association (log-odds).}
\item{se_outcome}{Standard error of SNP–Alzheimer association.}
}

REPRODUCIBILITY

The dataset can be regenerated with:

\preformatted{
educ <- extract_instruments(
  outcomes = "ieu-a-1239",
  p1 = 5e-08,
  clump = TRUE,
  r2 = 0.001,
  kb = 10000
)

alzheimer <- extract_outcome_data(
  snps = educ$SNP,
  outcomes = "ebi-a-GCST90027158",
  proxies = TRUE,
  rsq = 0.8
)

dat <- harmonise_data(
  exposure_dat = educ,
  outcome_dat = alzheimer,
  action = 2
)
}
}
\references{
Lee JJ et al. (2018) Nature Genetics.
Kunkle BW et al. (2019) Nature Genetics.
}
\keyword{GWAS}
\keyword{MendelianRandomization}
\keyword{datasets}
