% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_agreement.R
\name{fmi}
\alias{fmi}
\title{Fowlkes-Mallows Index (FMI)}
\usage{
fmi(part_A, part_B)
}
\arguments{
\item{part_A}{Integer vector length p}

\item{part_B}{Integer vector length p}
}
\value{
Scalar FMI
}
\description{
Geometric mean of pairwise precision and recall.
FMI = 1: perfect agreement; FMI = 0: no agreement.
}
