% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{I2}
\alias{I2}
\title{Auxiliary function to compute the I-squared heterogeneity index.}
\usage{
I2(y, s)
}
\arguments{
\item{y}{A list whose content must be checked.}

\item{s}{A list to use as a reference with which compare the first one.}
}
\value{
A length-one numeric vector.
}
\description{
\code{check_list_na()} compares two lists and fills in the missing
elements in the first with those included in the second. The
comparison is recursive in the sense that the process is repeated for
all lists included in those given.
}
\examples{
# [nothing for now]

}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
