% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{initialize,bayesmr_data-method}
\alias{initialize,bayesmr_data-method}
\alias{bayesmr_data-initialize}
\title{Initialize a \code{bayesmr_data} object.}
\usage{
\S4method{initialize}{bayesmr_data}(.Object, data = data.frame(), reorientation = TRUE)
}
\arguments{
\item{.Object}{Prototype object from class \code{\link[=bayesmr_data-class]{bayesmr_data}}.}

\item{data}{A data frame containing the columns \code{SNP}, \code{beta_exposure},
\code{beta_outcome}, \code{se_exposure}, and \code{se_outcome}. Dot variants of the four
numeric column names, namely \code{beta.exposure}, \code{beta.outcome}, \code{se.exposure},
and \code{se.outcome}, are also accepted and renamed internally.}

\item{reorientation}{A length-one logical value. If \code{TRUE}, rows with negative
\code{beta_exposure} are reoriented by changing the sign of both \code{beta_exposure}
and \code{beta_outcome}.}
}
\value{
An initialized \code{\link[=bayesmr_data-class]{bayesmr_data}} object.
}
\description{
Initialize a \code{bayesmr_data} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
