% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{initialize,bayesmr_fit_list-method}
\alias{initialize,bayesmr_fit_list-method}
\alias{bayesmr_fit_list-initialize}
\title{Initialize a \code{bayesmr_fit_list} object.}
\usage{
\S4method{initialize}{bayesmr_fit_list}(.Object, results = list())
}
\arguments{
\item{.Object}{Prototype object from class \code{\link[=bayesmr_fit_list-class]{bayesmr_fit_list}}.}

\item{results}{A list of fitted chain objects.}
}
\value{
An initialized \code{\link[=bayesmr_fit_list-class]{bayesmr_fit_list}} object.
}
\description{
Initialize a \code{bayesmr_fit_list} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
