% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_mcmc.R
\name{logpost_beta_util}
\alias{logpost_beta_util}
\title{Function to implement the Metropolis algorithm for an arbitrary posterior probability distribution}
\usage{
logpost_beta_util(beta, gamma, prior, data)
}
\arguments{
\item{data}{Integer value specifying the total the number of iterations
of the algorithm.}

\item{g}{A function defining the logarithm of the posterior density.}

\item{par}{A numeric value representing the starting value.}

\item{hpar}{A numeric value corresponding to the neighborhood where one looks
for a proposal value.}
}
\value{
A list with two components, \code{S} is a vector of the simulated draws,
and \code{accept_rate}, which gives the acceptance rate of the algorithm.
}
\description{
\code{metropolis()} implements a general Metropolis MCMC algorithm that can be applied to any
posterior probability provided by the user.
}
\examples{
data(bmi_sbp)

hpar <- list(mu_beta = 0,
             sigma2_beta = 0.1,
             psi2 = 0.3,
             tau2 = 0.4)

b <- seq(-0.8, 0.8, length.out = 100)
data <- bmi_sbp[, c("beta.exposure",
                    "beta.outcome",
                    "se.exposure",
                    "se.outcome")]
par <- list(gamma = 0.5)
res <- logpost_beta(b, par, hpar, data)
summary(res)

}
\seealso{
\code{\link{TODO}()} for computing ...
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
