% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_mcmc.R
\name{logpost_gamma}
\alias{logpost_gamma}
\title{Function to implement the Metropolis algorithm for an arbitrary posterior probability distribution}
\usage{
logpost_gamma(g, par, hpar, data)
}
\arguments{
\item{g}{A function defining the logarithm of the posterior density.}

\item{par}{A numeric value representing the starting value.}

\item{hpar}{A numeric value corresponding to the neighborhood where one looks
for a proposal value.}

\item{data}{Integer value specifying the total the number of iterations
of the algorithm.}
}
\value{
A list with two components, \code{S} is a vector of the simulated draws,
and \code{accept_rate}, which gives the acceptance rate of the algorithm.
}
\description{
\code{metropolis()} implements a general Metropolis MCMC algorithm that can be applied to any
posterior probability provided by the user.
}
\examples{
data(bmi_sbp)

hpar <- list(mu_gamma = 0,
             sigma2_gamma = 1,
             psi2 = 3,
             tau2 = 4)

g <- seq(-0.2, 0.2, length.out = 100)
data <- bmi_sbp[, c("beta.exposure",
                    "beta.outcome",
                    "se.exposure",
                    "se.outcome")]
par <- list(beta = 0.3)
res <- logpost_gamma(g, par, hpar, data)
summary(res)

}
\seealso{
\code{\link{TODO}()} for computing ...
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
