% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_mcmc.R
\name{metropolis}
\alias{metropolis}
\title{Function to implement the Metropolis algorithm for an arbitrary posterior probability distribution}
\usage{
metropolis(logpost, current, C, iter, ...)
}
\arguments{
\item{logpost}{A function defining the logarithm of the posterior density.}

\item{current}{A numeric value representing the starting value.}

\item{C}{A numeric value corresponding to the neighborhood where one looks
for a proposal value.}

\item{iter}{Integer value specifying the total the number of iterations
of the algorithm.}

\item{...}{Additional arguments to be passed to the \code{logpost}() function.}
}
\value{
A list with two components, \code{S} is a vector of the simulated draws,
and \code{accept_rate}, which gives the acceptance rate of the algorithm.
}
\description{
\code{metropolis()} implements a general Metropolis MCMC algorithm that can be applied to any
posterior probability provided by the user.
}
\examples{
lpost <- function(theta, s) {
  dcauchy(theta, s$loc, s$scale, log = TRUE) +
  dnorm(s$ybar, theta, s$se, log = TRUE)
}

data <- read.table("./data/buffalo_jan.txt", header = TRUE)
s <- list(loc = 10, scale = 2,
          ybar = mean(data$JAN),
          se = sd(data$JAN) / sqrt(nrow(data)))

set.seed(101)
out <- metropolis(lpost, 5, 20, 10000, s)

}
\seealso{
\code{\link{TODO}()} for computing ...
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
