% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{plot_posterior_K}
\alias{plot_posterior_K}
\title{Plot posterior distribution of K}
\usage{
plot_posterior_K(Kd, file = NULL)
}
\arguments{
\item{Kd}{Output from posterior_K()}

\item{file}{Optional file path}
}
\description{
Plot posterior distribution of K
}
