% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{plot_traces}
\alias{plot_traces}
\title{Trace plots for scalar parameters}
\usage{
plot_traces(out, file = NULL, pleiotropy = TRUE)
}
\arguments{
\item{out}{Object from remove_burnin()}

\item{file}{Optional file path}

\item{pleiotropy}{Declare whether the model assumes pleiotropy}
}
\description{
Trace plots for scalar parameters
}
