% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{posterior_K}
\alias{posterior_K}
\title{Posterior distribution of the number of active clusters K}
\usage{
posterior_K(out)
}
\arguments{
\item{out}{Object from remove_burnin()}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Data frame with columns k, count, proportion
}\if{html}{\out{</div>}}
}
\description{
Posterior distribution of the number of active clusters K
}
