% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{read_mcmc_output}
\alias{read_mcmc_output}
\title{Read all CSV output files written by mr_mixture.cpp}
\usage{
read_mcmc_output(res, burnin, pleiotropy = TRUE)
}
\arguments{
\item{res}{Output of the bayesmr_mix_het() function}

\item{burnin}{Number of initial iterations to discard}

\item{pleiotropy}{Declare whether the model assumes pleiotropy}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list with components described in the header above
}\if{html}{\out{</div>}}
}
\description{
Read all CSV output files written by mr_mixture.cpp
}
