% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{remove_burnin}
\alias{remove_burnin}
\title{Discard the first \code{burnin} iterations}
\usage{
remove_burnin(out, burnin, pleiotropy = TRUE)
}
\arguments{
\item{out}{Object from which to remove the burnin}

\item{burnin}{Number of initial iterations to discard}

\item{pleiotropy}{Declare whether the model assumes pleiotropy}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      Trimmed object with updated S
}\if{html}{\out{</div>}}
}
\description{
Discard the first \code{burnin} iterations
}
