% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnpmr_postprocess.R
\name{salso_partition}
\alias{salso_partition}
\title{Find the MCMC sample that minimises Binder's expected loss
relative to the PSM}
\usage{
salso_partition(out, loss = VI())
}
\arguments{
\item{out}{Object from remove_burnin()}

\item{loss}{salso loss function; default is VI() as in Dahl et al. (2022)
Use binder() for Binder loss with automatic a optimization}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ List with elements:
           $partition  integer vector length p
           $loss       expected loss of the selected partition
           $info       salso info data frame
}\if{html}{\out{</div>}}
}
\description{
Binder loss for a partition c is proportional to
\eqn{\sum_{i<j} { PSM_{i,j} \cdot 1(c_i \ne c_j)  +
              (1 - PSM_{i,j}) \cdot 1(c_i = c_j) }}
}
