% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_data-method}
\alias{show,bayesmr_data-method}
\alias{bayesmr_data-show}
\title{Show a \code{bayesmr_data} object.}
\usage{
\S4method{show}{bayesmr_data}(object)
}
\arguments{
\item{object}{An object of class \code{\link[=bayesmr_data-class]{bayesmr_data}}.}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of printing a
short description of the object.
}
\description{
Show a \code{bayesmr_data} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
