% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_full_fit_list-method}
\alias{show,bayesmr_full_fit_list-method}
\alias{bayesmr_full_fit_list-show}
\title{Show an instance of the \code{bayesmr_full_fit_list} class.}
\usage{
\S4method{show}{bayesmr_full_fit_list}(object)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_full_fit_list}}.}
}
\description{
Show an instance of the \code{bayesmr_full_fit_list} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
