% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_mix_fit_list-method}
\alias{show,bayesmr_mix_fit_list-method}
\alias{bayesmr_mix_fit_list-show}
\title{Show a \code{bayesmr_mix_fit_list} object.}
\usage{
\S4method{show}{bayesmr_mix_fit_list}(object)
}
\arguments{
\item{object}{An object of class \code{\link[=bayesmr_mix_fit_list-class]{bayesmr_mix_fit_list}}.}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of printing a
short description of the object.
}
\description{
Show a \code{bayesmr_mix_fit_list} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
