% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_model-method}
\alias{show,bayesmr_model-method}
\alias{bayesmr_model-show}
\title{Show an instance of the \code{bayesmr_model} class.}
\usage{
\S4method{show}{bayesmr_model}(object)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_model}}.}
}
\description{
Show an instance of the \code{bayesmr_model} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
