% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{subset,bayesmr_mix_het_fit_list-method}
\alias{subset,bayesmr_mix_het_fit_list-method}
\alias{bayesmr_mix_het_fit_list-subset}
\title{Subset MCMC parameters from a \code{bayesmr_mix_het_fit_list} object.}
\usage{
\S4method{subset}{bayesmr_mix_het_fit_list}(x, pars = character(), regex_pars = character(), ...)
}
\arguments{
\item{x}{An object of class \code{\link[=bayesmr_mix_het_fit_list-class]{bayesmr_mix_het_fit_list}}.}

\item{pars}{An optional character vector of exact parameter names to keep.}

\item{regex_pars}{An optional character vector of regular expressions used to
select parameter names.}

\item{...}{Further arguments passed to or ignored by the method.}
}
\value{
A \code{coda::mcmc.list} object containing the selected parameters.
}
\description{
Subset MCMC parameters from a \code{bayesmr_mix_het_fit_list} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
