% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{subset}
\alias{subset}
\title{Subset an object.}
\usage{
subset(x, ...)
}
\arguments{
\item{x}{An object to subset.}

\item{...}{Further arguments passed to methods.}
}
\value{
The return value depends on the method.
}
\description{
Generic function for subsetting BayesMR fitted objects by parameter names.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
