% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{summary,bayesmr_data-method}
\alias{summary,bayesmr_data-method}
\alias{bayesmr_data-summary}
\title{Summarize a \code{bayesmr_data} object.}
\usage{
\S4method{summary}{bayesmr_data}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[=bayesmr_data-class]{bayesmr_data}}.}

\item{...}{Further arguments passed to or ignored by the method.}
}
\value{
Invisibly returns \code{NULL}; called for its side effect of printing a
summary of the observed data.
}
\description{
Summarize a \code{bayesmr_data} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
