% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{summary,bayesmr_full_fit-method}
\alias{summary,bayesmr_full_fit-method}
\alias{bayesmr_full_fit-summary}
\title{Provide a summary of a \code{bayesmr_full_fit} class instance.}
\usage{
\S4method{summary}{bayesmr_full_fit}(object, include.burnin = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_full_fit}}.}

\item{include.burnin}{A length-one logical vector. If \code{TRUE} the
burnin iterations (if available) are included in the summary.}

\item{...}{Further arguments to pass on (currently ignored).}

\item{summary.Z}{A length-one logical vector. If \code{TRUE} the summary
also includes the latent configuration coordinates.}
}
\description{
Provide a summary of a \code{bayesmr_full_fit} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
