% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{summary,bayesmr_mix_fit_list-method}
\alias{summary,bayesmr_mix_fit_list-method}
\alias{bayesmr_mix_fit_list-summary}
\title{Summarize a \code{bayesmr_mix_fit_list} object.}
\usage{
\S4method{summary}{bayesmr_mix_fit_list}(object, include.burnin = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link[=bayesmr_mix_fit_list-class]{bayesmr_mix_fit_list}}.}

\item{include.burnin}{A length-one logical value indicating whether burn-in
iterations should be included when converting fitted objects to \code{coda}
objects before summarizing.}

\item{...}{Further arguments passed to or ignored by the method.}
}
\value{
For fitted objects and fitted-object lists, the result of
\code{summary.mcmc} or \code{summary.mcmc.list} via the
corresponding \code{coda} method. For \code{\link[=bayesmr_data-class]{bayesmr_data}}, invisibly returns
\code{NULL} after printing a summary of the data.
}
\description{
Summarize a \code{bayesmr_mix_fit_list} object.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
