% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_data-class}
\alias{bayesmr_data-class}
\alias{bayesmr_data}
\title{An S4 class to represent the data to use in a BayesMR model.}
\description{
An S4 class to represent the data to use in a BayesMR model.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list whose elements are the dissimilarity matrices corresponding
to the judgments expressed by the \emph{S} subjects/raters. These matrices
must be defined as a \code{dist} object.}

\item{\code{n}}{A length-one character vector representing the number of observations.}
}}

\examples{
showClass("bayesmr_data")

}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2024), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
