% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_fit_list-class}
\alias{bayesmr_fit_list-class}
\alias{bayesmr_fit_list}
\title{An S4 class to represent the results of fitting BayesMR model.}
\description{
An S4 class to represent the results of fitting BayesMR model using multiple
Markov Chain Monte Carlo chains.
}
\section{Slots}{

\describe{
\item{\code{results}}{An object of class \code{list}; list of \code{bayesmr_fit}
objects corresponding to the parallel MCMC chains simulated during the
estimation.}
}}

\examples{
showClass("bayesmr_fit_list")

}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2024), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\seealso{
\code{\link{bayesmr_fit}} for more details on the components of each element of
the list.
}
