% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{bayesmr_fit_to_mcmc}
\alias{bayesmr_fit_to_mcmc}
\title{Conversion of an \code{bayesmr_fit} object to an object of class \code{mcmc}.}
\usage{
bayesmr_fit_to_mcmc(res, include.burnin = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{An object of type \code{bayesmr_fit}.}

\item{include.burnin}{A logical scalar. If \code{TRUE} the burnin
iterations (if available) are not removed.}

\item{verbose}{A logical scalar. If \code{TRUE} prints additional
warnings during the conversion.}
}
\value{
An object of type \code{mcmc}.
}
\description{
\code{bayesmr_fit_to_mcmc} converts an object of class \code{bayesmr_fit}
to an object with class \code{mcmc}.
}
\examples{
\dontrun{
data(simdiss, package = "bayesmr")

G <- 3
p <- 2
prm.prop <- list(z = 1.5, alpha = .75)
burnin <- 2000
nsim <- 1000
seed <- 2301

set.seed(seed)

control <- list(burnin = burnin, nsim = nsim, z.prop = prm.prop[["z"]],
  alpha.prop = prm.prop[["alpha"]], verbose = TRUE)
sim.bayesmr <- bayesmr(simdiss, p, G, control)
sim.mcmc <- bayesmr_fit_to_mcmc(sim.bayesmr@results[[1]], TRUE)
plot(sim.mcmc)
}
}
\seealso{
\code{\link{bayesmr}()} for for fitting a DMBC model;
\code{\link{bayesmr_fit-class}};
\code{\link[coda]{mcmc}}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
