% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_init.R
\name{bayesmr_init}
\alias{bayesmr_init}
\title{Function to compute the starting values before fitting a BayesMR models.}
\usage{
bayesmr_init(data, p, G, random.start, partition)
}
\arguments{
\item{data}{A list whose elements are the dissimilarity matrices corresponding
to the judgments expressed by the \emph{S} subjects/raters. These matrices
must be defined as a \code{dist} object.}

\item{p}{A length-one numeric vector indicating the number of dimensions of the
latent space.}

\item{G}{A length-one numeric vector indicating the number of cluster to
partition the \emph{S} subjects.}

\item{random.start}{A length-one logical vector. If \code{TRUE} the starting
values are drawn randomly, otherwise.}

\item{partition}{A length-one numeric vector providing the user-defined
starting partition.}
}
\value{
A named \code{list} with the following items:
\describe{
\item{\code{z}: }{array of latent coordinates starting values}
\item{\code{x}: }{numeric vector of initial cluster memberships}
\item{\code{ng}: }{numeric vector of initial cluster sizes}
\item{\code{alpha}: }{numeric vector of alpha starting values}
\item{\code{eta}: }{numeric vector of eta starting values}
\item{\code{sigma2}: }{numeric vector of sigma2 starting values}
\item{\code{lambda}: }{numeric vector of lambda starting values}
}
}
\description{
\code{bayesmr_init()} is the main function that estimates a BayesMR model.
}
\examples{
data(simdiss, package = "bayesmr")
bayesmr_init(simdiss@diss, p = 2, G = 3, random.start = TRUE)
}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2026), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\seealso{
\code{\link{bayesmr}()} for fitting a BayesMR model.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
