% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_fit.R
\name{bayesmr_logLik}
\alias{bayesmr_logLik}
\title{Log-likelihood for BayesMR models.}
\usage{
bayesmr_logLik(data, gammaj, Gammaj)
}
\arguments{
\item{data}{A data frame ...}

\item{gammaj}{A numeric vector ...}

\item{Gammaj}{A numeric vector ...}
}
\value{
A length-one numeric vector of the log-likelihood value.
}
\description{
\code{bayesmr_logLik()} computes the log-likelihood value for a BayesMR model.
}
\references{
Venturini, S., Piccarreta, R. (2021), "A Bayesian Approach for Model-Based
Clustering of Several Binary Dissimilarity Matrices: the \pkg{bayesmr}
Package in \code{R}", Journal of Statistical Software, 100, 16, 1--35, <10.18637/jss.v100.i16>.
}
\seealso{
\code{\link{bayesmr}()}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
