% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointpost.R
\name{bayesmr_logLik}
\alias{bayesmr_logLik}
\title{Log-likelihood for BayesMR models.}
\usage{
bayesmr_logLik(gamma, beta, data, prior, log = TRUE)
}
\arguments{
\item{gamma}{A length-one numeric vector containing the gamma parameter value}

\item{beta}{A length-one numeric vector containing the beta parameter value}

\item{data}{An object of class \code{\link{bayesmr_data}}).}

\item{prior}{A list containing the prior hyperparameters. See
\code{\link{bayesmr_prior}()} for more details.}
}
\value{
A numeric matrix of the log-likelihood values.
}
\description{
\code{bayesmr_logLik()} computes the log-likelihood value for a BayesMR model.
}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2026), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
\seealso{
\code{\link{bayesmr}()}.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
