% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\docType{class}
\name{bayesmr_model-class}
\alias{bayesmr_model-class}
\alias{bayesmr_model}
\title{An S4 class to represent a BayesMR model.}
\description{
An S4 class to represent a BayesMR model.
}
\section{Slots}{

\describe{
\item{\code{p}}{A length-one character vector representing the number of dimensions
of the latent space to use in the MDS analysis.}

\item{\code{G}}{A length-one numeric vector representing the number of clusters to
partition the subjects into.}
}}

\examples{
showClass("bayesmr_model")

}
\references{
Consonni, G., Venturini, S., Castelletti, F. (2024), "Bayesian Hierarchical Modeling for
Two-Sample Summary-Data Mendelian Randomization under Heterogeneity, working paper.
}
