% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_init.R
\name{bayesmr_startvalues}
\alias{bayesmr_startvalues}
\alias{startvalues_bayesmr}
\alias{check_startvalues}
\title{Auxiliary Function for Setting BayesMR Model Starting Values}
\usage{
bayesmr_startvalues(gamma, beta, ...)

startvalues_bayesmr(gamma, beta, ...)

check_startvalues(startvalues)
}
\arguments{
\item{gamma}{A named list containing the hyperparameters for the prior
distribution of the \eqn{\eta_1,\ldots,\eta_G} parameters. It should
contain two numeric vectors, namely \code{a} and \code{b}.}

\item{beta}{A named list containing the hyperparameters for the prior
distributions of the \eqn{\sigma^2_1,\ldots,\sigma^2_G} parameters. It
should contain two numeric scalars, namely \code{a} and \code{b}.}

\item{gammaj}{A named list containing the hyperparameters for the prior
distribution of the \eqn{\eta_1,\ldots,\eta_G} parameters. It should
contain two numeric vectors, namely \code{a} and \code{b}.}

\item{Gammaj}{A named list containing the hyperparameters for the prior
distribution of the \eqn{\eta_1,\ldots,\eta_G} parameters. It should
contain two numeric vectors, namely \code{a} and \code{b}.}

\item{prior}{A named list of prior hyperparameters.}
}
\value{
A list with the prior hyperparameters as components.
}
\description{
{
\code{bayesmr_startvalues()} is an auxiliary function as user interface for
\code{bayesmr()} fitting. Typically only used when calling the \code{bayesmr()}
function. It is used to set prior hyperparameters.

\code{prior_bayesmr()} is an alias for \code{bayesmr_startvalues()}.

\code{check_prior()} is an auxiliary function that verifies the
correctness of the prior hyperparameters provided before a BayesMR is fitted
with \code{\link{bayesmr}()}.

\code{update_prior()} is an auxiliary function to modify a set of prior
choices using a new value of \emph{p} and \emph{G}. It is intended for
internal use mainly in the \code{\link{bayesmr_ic}()} function.
}
}
\examples{
\dontrun{
data(simdiss, package = "bayesmr")
# Shorter run than default.
sim.fit <- bayesmr(simdiss,
  control = bayesmr_control(burnin = 1000, nsim = 2000, thin = 1, verbose = TRUE),
  prior = bayesmr_startvalues(gamma = list(mean = 0, var = 1)))
}

}
\seealso{
\code{\link{bayesmr}()}
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
\keyword{based}
\keyword{clustering}
\keyword{model}
