% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\name{BayesMRclus-package}
\alias{BayesMRclus-package}
\alias{BayesMRclus-pkg}
\alias{BayesMRclus-p}
\alias{BayesMRclus}
\title{BayesMRclus: a package for Bayesian inference in two-sample summary data Mendelian Randomization designs
under heterogeneity}
\description{
The \pkg{BayesMRclus} package contains functions for performing Bayesian
inference in two-sample summary data Mendelian Randomization designs under
heterogeneity.

The main functions included in the package are:
\itemize{
\item \code{\link{TODO}()}
}
}
\details{
_PACKAGE
}
\keyword{internal}
