% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{initialize,bayesmr_het_fit_list-method}
\alias{initialize,bayesmr_het_fit_list-method}
\alias{bayesmr_het_fit_list-initialize}
\title{Create an instance of the \code{bayesmr_het_fit_list} class using new/initialize.}
\usage{
\S4method{initialize}{bayesmr_het_fit_list}(.Object, results = list())
}
\arguments{
\item{.Object}{Prototype object from the class \code{\link{bayesmr_het_fit_list}}.}

\item{results}{A list whose elements are the \code{bayesmr_het_fit} objects for
each simulated chain.}
}
\description{
Create an instance of the \code{bayesmr_het_fit_list} class using new/initialize.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
