% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_data-method}
\alias{show,bayesmr_data-method}
\alias{bayesmr_data-show}
\title{Show an instance of the \code{bayesmr_data} class.}
\usage{
\S4method{show}{bayesmr_data}(object)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_data}}.}
}
\description{
Show an instance of the \code{bayesmr_data} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
