% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{show,bayesmr_fit-method}
\alias{show,bayesmr_fit-method}
\alias{bayesmr_fit-show}
\title{Show an instance of the \code{bayesmr_fit} class.}
\usage{
\S4method{show}{bayesmr_fit}(object)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_fit}}.}
}
\description{
Show an instance of the \code{bayesmr_fit} class.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
