% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmrclus_classes.R
\name{summary,bayesmr_data-method}
\alias{summary,bayesmr_data-method}
\alias{bayesmr_data-summary}
\title{Provide a summary of a \code{bayesmr_data} class instance.}
\usage{
\S4method{summary}{bayesmr_data}(object)
}
\arguments{
\item{object}{An object of class \code{\link{bayesmr_data}}.}
}
\description{
Provide a summary of a \code{bayesmr_data} class instance.
}
\author{
Sergio Venturini \email{sergio.venturini@unicatt.it}
}
